Primeiras modificações da nova versão da refatoração do código
