+++ src/contexts/AuthContext.tsx
@@ -58,7 +58,7 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
 
     api.defaults.headers['Authorization'] = `Bearer ${data.token}`
 
-    Router.push('/')
+    Router.push(`/${data.user.id}`)
   }
   
   function logOut() {
