+++ src/pages/[id]/index.tsx
@@ -1,0 +1,35 @@
+import { GetServerSideProps } from "next";
+import { parseCookies } from "nookies";
+import Feed from "../../components/Feed/Feed";
+import Header from "../../components/Header/Header";
+import { useAuthContext } from "../../contexts/AuthContext";
+
+export default function AppPage({ urlParams }: any) {
+  const { user } = useAuthContext();
+
+  return (
+    <div className="w-full h-full flex flex-col justify-center px-14 mt-4 mb-10 gap-6">
+      <Header user={user} />
+      <Feed params={urlParams} />
+    </div>
+  );
+}
+
+export const getServerSideProps: GetServerSideProps = async (ctx) => {
+  const { ["nextauth.token"]: token } = parseCookies(ctx);
+  
+  if (!token) {
+    return {
+      redirect: {
+        destination: "/sign-in",
+        permanent: false,
+      },
+    };
+  }
+
+  return {
+    props: {
+      urlParams: ctx.query
+    },
+  };
+};
